Move test_utils to top level
