+++ p2panda-auth/src/lib.rs
@@ -4,5 +4,7 @@ pub mod group;
 pub mod traits;
 mod access;
 pub mod graph;
+#[cfg(any(test, feature = "test_utils"))]
+mod test_utils;
 
 pub use access::{Access, AccessLevel};
